Implement #inorder, #left_child?, #right_child?
The #inorder method uses recursion to traverse the tree in order. It
does so by visiting the left child, root, and right child.

The #left_child? and #right_child? methods were created to make it
easier to find out if a node has children or not as they return boolean
values.